Add option to configure default environment
The default environment that will be sourced is normally the ORACLE_HOME of the active Grid Infrastructure. Or none, if there is no GI installed. This commit allows to configure a different environment by specifying the variable "GV_DEFAULT_ENV_NAME" in ~/.ocenvrc. The value in this variable must be the name of one of the environment-aliases for the ORACLE_HOMEs or instances on the system.